Перенос сотрудников в проект доменной зоны
