+++ WebStore/Data/TestData.cs
@@ -1,4 +1,4 @@
-﻿using WebStore.Models;
+﻿using WebStore.Domain.Entities;
 
 namespace WebStore.Data
 {
