+++ WebStore/WebStore.csproj
@@ -189,6 +189,14 @@
 		<PackageReference Include="QRCoder" Version="1.4.3" />
 	</ItemGroup>
 
+	<ItemGroup>
+	  <ProjectReference Include="..\WebStore.Domain\WebStore.Domain.csproj" />
+	</ItemGroup>
+
+	<ItemGroup>
+	  <Folder Include="Models\" />
+	</ItemGroup>
+
 	<ProjectExtensions>
 		<VisualStudio>
 			<UserProperties appsettings_1json__JsonSchema="" properties_4launchsettings_1json__JsonSchema="" />
